Minor refactor:
Moved basic entities from UI to core module
