+++ hermescore/src/main/java/org/hermes/core/entities/Tick.java
@@ -1,4 +1,4 @@
-package org.hermes.gui.javafx.orderbook.model;
+package org.hermes.core.entities;
 
 public class Tick {
     private final String ticker;
