+++ hermescore/src/main/java/org/hermes/core/entities/TickType.java
@@ -1,0 +1,6 @@
+package org.hermes.core.entities;
+
+public enum TickType {
+    BID,
+    ASK
+}
