--- hermesui/src/main/java/org/hermes/gui/javafx/orderbook/model/TickType.java
@@ -1,6 +1,0 @@
-package org.hermes.gui.javafx.orderbook.model;
-
-public enum TickType {
-    BID,
-    ASK
-}
